Fix -Wshorten-64-to-32 on x64
- Enable only on 64 bit platforms
- Suppress dawn/native backend code

Bug: 523358231
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/327015
Commit-Queue: Shrek Shao <shrekshao@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>

+++ src/dawn/common/GPUInfo.cpp
@@ -64,7 +64,7 @@ const uint16_t& DriverVersion::operator[](size_t i) const {
     return mDriverVersion.operator[](i);
 }
 
-uint32_t DriverVersion::size() const {
+size_t DriverVersion::size() const {
     return mDriverVersion.size();
 }
